desktop: get scrollToIndex working
